Issue #983: extract provider crate
