+++ crates/tau-coding-agent/Cargo.toml
@@ -29,6 +29,7 @@ tau-runtime = { path = "../tau-runtime" }
 tau-events = { path = "../tau-events" }
 tau-deployment = { path = "../tau-deployment" }
 tau-startup = { path = "../tau-startup" }
+tau-provider = { path = "../tau-provider" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
