+++ crates/tau-provider/src/lib.rs
@@ -1,0 +1,21 @@
+mod auth;
+mod claude_cli_client;
+mod cli_executable;
+mod client;
+mod codex_cli_client;
+mod credential_store;
+mod credentials;
+mod fallback;
+mod gemini_cli_client;
+mod types;
+
+pub use auth::*;
+pub use claude_cli_client::*;
+pub use cli_executable::is_executable_available;
+pub use client::*;
+pub use codex_cli_client::*;
+pub use credential_store::*;
+pub use credentials::*;
+pub use fallback::*;
+pub use gemini_cli_client::*;
+pub use types::*;
